- Created a partial class just for the update method
